change mapper location from restController to service and update Services and RestControllers
