--- src/main/java/com/example/search/model/Distributor.java
@@ -1,22 +1,0 @@
-package com.example.search.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.OneToOne;
-
-@Entity
-@Getter
-@Setter
-public class Distributor {
-    @Id
-    private String id;
-
-    private String name;
-
-    @OneToOne
-    private Address address;
-
-}
